feat: Add applyService and acceptServiceRequest functions
This commit adds the applyService and acceptServiceRequest functions to the serviceFunc module. The applyService function sends a POST request to the server to apply for a service, while the acceptServiceRequest function sends a POST request to accept a service request. These functions enable users to apply for services and accept service requests, respectively.