feat: Add authorCreated field to formService in ServiceProvider
The code changes in `ServiceProvider.jsx` add an `authorCreated` field to the `formService` object. This field is set to the `currentUser.uid` value when there is a logged-in user. This change ensures that the author of a service is correctly recorded when creating a new service.

Note: This message has been generated based on the provided code changes and recent commits.